requirements.txt
###### Requirements without Version Specifiers ######
numpy
#scipy
#sympy
pandas
matplotlib
#plotly
scikit-learn
db-sqlite3
#seaborn
#pubchempy
biopython
bioservices
#qgrid
#kaleido
#modeller
py3dmol
rdkit-pypi 
prolif
